move keys endpoint to developer and created get scopes endpoint
