+++ app/router/service_provider.py
@@ -1,41 +1,16 @@
 from fastapi import APIRouter, Depends, HTTPException, status, Request, Query
 from sqlalchemy import desc
 from sqlalchemy.orm import Session
-from typing import List
 from urllib.parse import quote
 
 from app.models import ServiceProvider, User, ClientScope, Scope
-from app.schemas import ServiceProviderSchema, SessionSchema, GetServiceProviderDetailsSchema
+from app.schemas import ServiceProviderSchema, SessionSchema
 from app.config import Settings
 from app.database import get_db
 from app.utils import generate_authorization_code, verify_session, get_current_user
 from fastapi.responses import RedirectResponse, JSONResponse
 
 router = APIRouter()
-
-
-@router.get("/credentials", response_model=List[GetServiceProviderDetailsSchema])
-def read_service_providers(current_user: User = Depends(get_current_user), db: Session = Depends(get_db)):
-    service_providers = db.query(ServiceProvider).filter(ServiceProvider.developer_id == current_user.id).order_by(ServiceProvider.created_at).all()
-    service_providers_with_scopes = []
-
-    for service_provider in service_providers:
-        client_scopes = (
-            db.query(Scope.scope)
-            .join(ClientScope, ClientScope.scope_id == Scope.id)
-            .filter(ClientScope.service_provider_id == service_provider.id)
-            .all()
-        )
-
-        scope_strings = [scope.scope for scope in client_scopes]
-
-        service_provider_data = {
-            **service_provider.__dict__,
-            "scopes": scope_strings       
-        }
-        service_providers_with_scopes.append(service_provider_data)
-
-    return service_providers_with_scopes
 
 
 @router.post("/create/")
