+++ app/schemas.py
@@ -22,7 +22,7 @@ class ServiceProviderSchema(BaseModel):
     scope: List[int]
 
 
-class GetServiceProviderDetailsSchema(BaseModel):
+class GetDeveloperKeysSchema(BaseModel):
     id: int
     name: str
     created_at: datetime
@@ -87,3 +87,7 @@ class GetDeveloperDetailsSchema(BaseModel):
     email: str
     phone_number: str
     is_verified: bool
+
+class GetAllScopesSchema(BaseModel):
+    id: int
+    scope: str
